+++ opentelemetry-sdk/src/trace/mod.rs
@@ -38,7 +38,7 @@ pub use in_memory_exporter::{InMemorySpanExporter, InMemorySpanExporterBuilder};
 pub use id_generator::{IdGenerator, RandomIdGenerator};
 pub use links::SpanLinks;
 pub use provider::{SdkTracerProvider, TracerProviderBuilder};
-pub use sampler::{Sampler, ShouldSample};
+pub use sampler::{Sampler, SamplingDecision, SamplingResult, ShouldSample};
 pub use span::Span;
 pub use span_limit::SpanLimits;
 pub use span_processor::{
@@ -60,13 +60,14 @@ mod runtime_tests;
 mod tests {
     use super::*;
     use crate::error::OTelSdkResult;
+    use crate::trace::{SamplingDecision, SamplingResult};
     use crate::{
         trace::span_limit::{DEFAULT_MAX_EVENT_PER_SPAN, DEFAULT_MAX_LINKS_PER_SPAN},
         trace::{InMemorySpanExporter, InMemorySpanExporterBuilder},
     };
     use opentelemetry::{
         baggage::BaggageExt,
-        trace::{SamplingDecision, SamplingResult, SpanKind, Status, TraceContextExt, TraceState},
+        trace::{SpanKind, Status, TraceContextExt, TraceState},
     };
     use opentelemetry::{testing::trace::TestSpan, InstrumentationScope};
     use opentelemetry::{
